Reorganize tests.
Condense all node-classification tests into single test module
one level up.
